+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(cs9example)
+
+test_check("cs9example")